Fix format specifier -Wformat warning
Fixes warnings like this:

    /local/mnt/workspace/vp/qemu/tests/tcg/hexagon/system/mmu.h:232:17: error: format specifies type 'unsigned long' but the argument has type 'uint32_t' (aka 'unsigned int') [-Werror,-Wformat]
      229 |         printf(" PA:0x%09llx SZ:%s (0x%lx)",
          |                                       ~~~
          |                                       %x
      230 |                 PA,
      231 |                 pgsize_str[hex_tlb_pgsize(entry)],
      232 |                 hex_tlb_page_size(entry));
          |                 ^~~~~~~~~~~~~~~~~~~~~~~~
    1 error generated.

+++ tests/tcg/hexagon/system/mmu.h
@@ -229,7 +229,7 @@ static void hex_dump_mmu_entry(uint64_t entry)
         printf(" PA:0x%09llx SZ:%s (0x%lx)",
                 PA,
                 pgsize_str[hex_tlb_pgsize(entry)],
-                hex_tlb_page_size(entry));
+                (unsigned long)hex_tlb_page_size(entry));
         printf("\n");
     } else {
         printf("invalid\n");
